# gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/resources/openapi/management-openapi-v2.yaml
openapi: 3.0.3
info:
    title: Gravitee.io APIM - Management API
    description: |-
        This is the OpenAPI specification for our new version of APIM Management API.
    contact:
        email: team-apim@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    version: 2.0.0
security:
    - BasicAuth: []
    - CookieAuth: []
servers:
    - url: /management/v2
      description: Gravitee.io APIM - Management API -v2
...
paths:
...
    /environments/{envId}/apis/{apiId}/analytics/logs:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
        get:
            parameters:
                - $ref: "#/components/parameters/pageParam"
                - $ref: "#/components/parameters/perPageParam"
            tags:
                - Analytics - Logs
            summary: Get API logs
            description: |-
                Get API logs.

                User must have the API_LOG[READ] permission.
            operationId: getApiLogs
            responses:
                "200":
                    $ref: "#/components/responses/ApiLogsResponse"
                default:
                    $ref: "#/components/responses/Error"

    # Plugins - Endpoints
...
components:
    schemas:
...
        Error:
            type: object
            properties:
                httpStatus:
                    type: integer
                    format: int32
                    description: The error code
                    example: 400
                message:
                    type: string
                    description: The error message
                    example: Bad request
                technicalCode:
                    type: string
                    description: A technical code to identify the error
                    example: invalid.import.definition
                parameters:
                    type: object
                    description: A map of parameters to be used in the error message
                    additionalProperties:
                        type: string
                details:
                    type: array
                    description: A list of details about the error
                    items:
                        type: object
                        properties:
                            message:
                                type: string
                                description: The error message
                                example: Bad request
                            location:
                                type: string
                                description: The json path of the field in error.
                                example: updateApi.properties[0].key
                            invalidValue:
                                type: object
                                description: The invalid value.
...
        Links:
            description: List of links for pagination
            properties:
                self:
                    type: string
                    description: Link to current resource
                first:
                    type: string
                    description: In a paginated response, link to the first page
                last:
                    type: string
                    description: In a paginated response, link to the last page
                previous:
                    type: string
                    description: In a paginated response, link to the previous page. Maybe null if current is the first page
                next:
                    type: string
                    description: In a paginated response, link to the next page. Maybe null if current is the last page
...
        Pagination:
            description: Generic object to handle pagination data.
            type: object
            properties:
                page:
                    type: integer
                    description: The current page.
                perPage:
                    type: integer
                    description: The number of items requested per page.
                pageCount:
                    type: integer
                    description: The total number of pages.
                pageItemsCount:
                    type: integer
                    description: The number of items for the current page.
                totalCount:
                    type: integer
                    description: The total number of items.
...
        BasePlan:
            type: object
            properties:
                id:
                    type: string
                    description: Plan's uuid.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                name:
                    type: string
                    description: Plan's name. Duplicate names can exists.
                    example: My Api plan
                description:
                    type: string
                    description: Plan's description. A short description of your plan.
                    example: I can use a hundred characters to describe this plan.
                apiId:
                    type: string
                    description: Id of the API owning the plan.
                    example: 6c530064-0b2c-4004-9300-640b2ce0047b
                security:
                    $ref: "#/components/schemas/PlanSecurity"
...
        PlanSecurity:
            type: object
            properties:
                type:
                    $ref: "#/components/schemas/PlanSecurityType"
                configuration:
                    type: object
            required:
                - type
        PlanSecurityType:
            type: string
            description: Plan security type.
            example: KEY_LESS
            enum:
                - KEY_LESS
                - API_KEY
                - OAUTH2
                - JWT
...
        BaseApplication:
            type: object
            properties:
                id:
                    type: string
                    description: Application's uuid.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                name:
                    type: string
                    description: Application's name. Duplicate names can exist.
                    example: My App
                description:
                    type: string
                    description: Application's description. A short description of your App.
                    example: I can use a hundred characters to describe this App.
                domain:
                    type: string
                    description: Application's domain. Domain used by application, if relevant.
                    example: https://my-app.com
                type:
                    type: string
                    description: Application's type.
                    example: iOS
                primaryOwner:
                    $ref: "#/components/schemas/PrimaryOwner"
...
        PrimaryOwner:
            type: object
            properties:
                id:
                    type: string
                    description: Owner's uuid.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                email:
                    type: string
                    description: Owner's email. Can be null if owner is a group.
                displayName:
                    type: string
                    description: Owner's name.
                    example: John Doe
                    minLength: 1
                type:
                    $ref: "#/components/schemas/MembershipMemberType"
...
        MembershipMemberType:
            type: string
            description: The type of membership
            enum:
                - USER
                - GROUP
...
        ApiLog:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    description: The date (as timestamp) of the log.
                    example: 1581256457163
                id:
                    type: string
                    description: The internal UUID of the Log.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                requestId:
                    type: string
                    description: The id of the request.
                    example: 1719d8d1-8d01-48f6-99d8-d18d0178f6d4
                clientIdentifier:
                    type: string
                    description: The client identifier of the request.
                    example: 12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0
                plan:
                    $ref: "#/components/schemas/BasePlan"
                application:
                    $ref: "#/components/schemas/BaseApplication"
                transactionId:
                    type: string
                    description: The id of the transaction.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                status:
                    type: integer
                    description: The response's status.
                    example: 200
                requestEnded:
                    type: boolean
                    description: The flag indicating if the request has ended.
                    example: true

        # Documentation pages
...
    parameters:
...
        envIdParam:
            name: envId
            in: path
            required: true
            description: Id or Hrid (Human readable Id) of an environment.
            schema:
                type: string
                default: DEFAULT
        apiIdParam:
            name: apiId
            in: path
            required: true
            description: Id of an API.
            schema:
                type: string
...
        pageParam:
            name: page
            in: query
            required: false
            description: The page number for pagination.
            schema:
                type: integer
                default: 1
        perPageParam:
            name: perPage
            in: query
            required: false
            description: |
                The number of items per page for pagination.
            schema:
                type: integer
                default: 10
        ###############################
        # Other Query Parameters      #
        ###############################
...
    responses:
...
        Error:
            description: Generic error response
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"

        # Analytics - Logs
        ApiLogsResponse:
            description: Page of API Logs
            content:
                application/json:
                    schema:
                        title: "ApiLogsResponse"
                        properties:
                            data:
                                description: List of API logs.
                                type: array
                                items:
                                    $ref: "#/components/schemas/ApiLog"
                            pagination:
                                $ref: "#/components/schemas/Pagination"
                            links:
                                $ref: "#/components/schemas/Links"
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
        CookieAuth:
            type: apiKey
            in: cookie
            name: Auth-Graviteeio-APIM